Extract a base class for Assert and Require
